Got streaming working.
For those playing at home the main problem I had was "invalid
hexadecimal in chunked encoding" - I think that was happening when I
switched to using `streaming: true` without returning something with an
each method as the body.

The second tricky issue was streaming media formats - the PSP requires
Content-Length be returned in order to stream. When I stopped removing
it from the headers on unmodified responses, it suddenly started
working. Woop.